Add the DatabaseHelper class
The DatabaseHelper extends SQLiteOpenHelper and provides the ability
to create and open the FlagsQuiz database. Updating the database schema
is currently not supported

The database schema is stores in the res/raw/schema.sql.

DatabaseHelper uses IOUtils from Apache Commons IO so I added this
dependency to the build.gradle of the app module. Also, I removed
the testing dependencies as they are not needed.