Update Gradle plugin to v4.1 and Java to v8
Now that Android Studio 3.0 is out, Gradle 4.1 and Java 8 are supported.

In order to properly build this project, it is necessary to sync
Gradle and clean it first.

Also, when starting up the emulator, I experienced this error:
Emulator: qemu-system-i386.exe: goldfish_battery_read: Bad offset
This error is annoying, but it doesn't really interfere with working
on the project.