Add region service
Created interface RegionService and its implementation,
RegionServiceImpl. It can only be used to get data as saving regions
is not needed in the app.

Also, updated the toString() methods in entity classes to include id.

Dagger module and component now don't include a string, which was used
for testing. MainActivity now has RegionService injected instead of
a string, again for testing only.