Add the preferences resource file
Now the preferences are specified in the res/xml/preferences.xml file.
The regions are a dynamic list, and they should be populated using the
PreferencesService#populateRegions().

Also, I changed the PreferencesService. It doesn't need the set method
anymore because the preferences can be saved automatically in the
PreferencesFragment.

Furthermore, I enabled dependency injection for Fragments.